Window: fullscreen mode: remove no longer needed code
That code was added in:
commit 8f1324a9f72a0bbbf9fcdd1e6b8578dfb3c648e6
"Hide the fullscreen control when you change from workspace or when you
minimize gedit".

In those cases, there is now the "leave-notify-event" signal that is
emitted on the fullscreen_eventbox. GeditWindow listens to that signal,
it calls the on_fullscreen_eventbox_leave_notify_event() function, which
hides the GtkRevealer when needed.

This fixes the following problem: when in fullscreen mode we click on
the open recent menu button, the menu was shown but the headerbar was
hidden. This is now fixed, the headerbar (and revealer) is still shown
when the open recent menu is shown.