build: list of deps: simplify
It's as clear when providing the version requirement inside the
dependency() function call. It takes less lines. And each time that a
variable is declared in meson, we don't know if it's a "global" variable
or not (i.e. we don't know if that variable is used in another
meson.build file).