build: simplify handling of gtk_doc option
It's not a good practice to have automatic build options: a build option
that is enabled or disabled depending on whether the dependency is
installed or not. It makes the build less predictable.

In this case when enabling the gtk_doc option, it was actually disabled
if gtk-doc was not installed. It's better in that case to show an error.

And now if the gtk_doc option is enabled, and gtk-doc is not installed,
the error shown is:
ERROR: Program(s) ['gtkdoc-scan'] not found or not executable

I think the error message is clear enough, no need to add a dependency()
call. It also simplifies the build instructions.