build: avoid the use of dir variables
All those *dir variables make the code a bit less clear in my opinion.
We need to look where the variable is assigned for the last time. All
variables in meson are global, and they are not const, so another
meson.build file can overwrite the value by accident.

When reading "get_option('prefix')" etc, everybody who has learned meson
is able to understand the code without looking elsewhere in the build
system. And it's short enough.